# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
      - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
      - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
              properties:
                user_id:
                  type: integer
                  format: uint
              required:
              - user_id
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    patch:
      summary: Update a task partially
      tags:
      - tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint
      requestBody:
        description: Fields to update in the task
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task
      tags:
      - tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint
      responses:
        '204':
          description: Task successfully deleted
  /users:
    get:
      summary: Get all users
      tags:
      - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
      - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    patch:
      summary: Update a user partially
      tags:
      - users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint
      requestBody:
        description: Fields to update in the user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete a user
      tags:
      - users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint
      responses:
        '204':
          description: User successfully deleted
  /tasks/user/{user_id}:
    get:
      summary: Get tasks by user ID
      tags:
      - tasks
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          format: uint
      responses:
        '200':
          description: A list of tasks for the given user ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string
        deleted_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time